Move test packages also
