+++ plugins/cloud-azure/src/test/java/org/elasticsearch/cloud/azure/ESBaseDirectoryTestCase.java
@@ -1,4 +1,4 @@
-package org.apache.lucene.store;
+package org.elasticsearch.cloud.azure;
 
 /*
  * Licensed to Elasticsearch under one or more contributor
@@ -19,6 +19,7 @@ package org.apache.lucene.store;
  * under the License.
  */
 
+import org.apache.lucene.store.BaseDirectoryTestCase;
 import org.apache.lucene.util.LuceneTestCase;
 import org.apache.lucene.util.TimeUnits;
 import org.elasticsearch.bootstrap.BootstrapForTesting;
